テキストファイルからラベルを生成できるようにした

+++ make_htsvoice.sh
@@ -1,16 +1,25 @@
 #!/bin/bash
 
 if [ $# -lt 2 ]; then
-  echo "Usage: $0 [-G APIKey] InputDirPath HTSVoicePath [OpenJTalkOption] [BaseHTSVoicePath]"
+  echo ""
+  echo "Usage: $0 [-G APIKey] [-T] InputDirPath HTSVoicePath [OpenJTalkOption] [BaseHTSVoicePath]"
+  echo ""
+  echo "-G APIKey : Specify API key only when using google cloud speech API"
+  echo "-T        : Create a label from a text file with the same name as the audio file in InputDirPath"
+  echo "InputDirPath (Required) : Folder with audio files used for training"
+  echo "HTSVoicePath (Required) : File path of htsvoice file to output"
+  echo "OpenJTalkOption  : Arguments to pass to Open JTalk used when creating the label"
+  echo "BaseHTSVoicePath : File path of the acoustic model to pass to Open JTalk used when creating the label"
+  echo ""
   exit -1
 fi
 
-APIMODE="J" # J:julius, G:google cloud speech API
+APIMODE="J" # J:julius, T:textfile, G:google cloud speech API
 APIKEY=""
 OPTIONS=""
 
 UPPERF0=500 # 基本周波数抽出の上限 (Hz) (女性の声の場合は 500 (Hz)・男性の声の場合は 200 (Hz) 程度が良いらしい)
-NITER=50 # トレーニングの反復回数 (増やすと精度が上がるらしいが、その分時間がかかる)
+NITER=10 # トレーニングの反復回数 (増やすと精度が上がるらしいが、その分時間がかかる)
 
 STARTPATH=`pwd`
 
@@ -25,6 +34,13 @@ if [ $1 = "-G" ]; then
   HTSVOICEPATH="$(cd $(dirname "$4") && pwd)/$(basename "$4")"
   cd "$STARTPATH"
   shift 4
+elif [ $1 = "-T" ]; then
+  APIMODE="T"
+  INPUTDIRPATH=$(cd "$2" && pwd)
+  cd "$STARTPATH"
+  HTSVOICEPATH="$(cd $(dirname "$3") && pwd)/$(basename "$3")"
+  cd "$STARTPATH"
+  shift 3
 else
   INPUTDIRPATH=$(cd "$1" && pwd)
   cd "$STARTPATH"
@@ -33,7 +49,7 @@ else
   shift 2
 fi
 
-# OpenJTalk での音声認識時に使用する音響モデル
+# OpenJTalk でのラベル作成時に使用する音響モデル
 mFlag=0
 while [ $# -ge 2 -a "$(echo "$1" | cut -c 1)" = "-" ]
 do
@@ -44,8 +60,10 @@ do
   shift 2
 done
 if [ $# -lt 1 ]; then
+  # デフォルトの音響モデルを利用
   OPTIONS=" ${OPTIONS} -m /usr/share/hts-voice/nitech-jp-atr503-m001/nitech_jp_atr503_m001.htsvoice "
 elif [ ${mFlag} -eq 0 ]; then
+  # 指定された音響モデルを利用
   OPTIONS=" ${OPTIONS} -m \"$(cd $(dirname "$1") && pwd)/$(basename "$1")\" "
 fi
 cd "$STARTPATH"
@@ -70,35 +88,61 @@ rm -rf tmp
 mkdir tmp
 for file in `ls "$INPUTDIRPATH"`
 do
-  echo -e "Convert ${file} to ${file}.raw"
+  # テキストファイルを除外
+  if [ "${file##*.}" != 'txt' ]; then
 
-  # ffmpeg で wav に変換しておく
-  ffmpeg -i ${INPUTDIRPATH}/${file} -acodec pcm_s16le -ar 44100 "${INPUTDIRPATH}/${file%.*}.wav" > /dev/null 2>&1
+    echo -e "Convert ${file} to ${file}.raw"
 
-  # 音量を計測
-  vol=`sox "${INPUTDIRPATH}/${file%.*}.wav" -n stat -v 2>&1`
-  volGain=`echo "scale=9; ${vol} / 2.86" | bc`
+    # ffmpeg で wav に変換しておく
+    ffmpeg -i ${INPUTDIRPATH}/${file} -acodec pcm_s16le -ar 44100 "${INPUTDIRPATH}/${file%.*}.wav" > /dev/null 2>&1
 
-  # raw ファイルを作成
-  sox "${INPUTDIRPATH}/${file%.*}.wav" -t raw -r 16k -e signed-integer -b 16 -c 1 -B "$(pwd)/tmp/${file}.raw" vol ${volGain}
+    # 音量を計測
+    vol=`sox "${INPUTDIRPATH}/${file%.*}.wav" -n stat -v 2>&1`
+    volGain=`echo "scale=9; ${vol} / 2.86" | bc`
 
-  # wav ファイルを削除
-  rm "${INPUTDIRPATH}/${file%.*}.wav"
+    # raw ファイルを作成
+    sox "${INPUTDIRPATH}/${file%.*}.wav" -t raw -r 16k -e signed-integer -b 16 -c 1 -B "$(pwd)/tmp/${file%.*}.raw" vol ${volGain}
+
+    # wav ファイルを削除
+    rm "${INPUTDIRPATH}/${file%.*}.wav"
+
+  fi
 done
 
+# make_label_fromtext.py でラベルデータを作成
+if [ ${APIMODE} = "T" ] ; then
+
+  # raw ファイルをコピー
+  cp tmp/*.raw $STARTPATH/tools/segment_adapt/voices/
+
+  # make_label_fromtext.py でテキストファイルからラベルデータを作成
+  echo -e "\nExecute make_label_fromtext.py ...\n"
+  python3 "$STARTPATH/make_label_fromtext.py" ${INPUTDIRPATH} "${OPTIONS}"
+
+  # segment_adapt.pl を実行
+  cd $STARTPATH/tools/segment_adapt/
+  perl "$STARTPATH/tools/segment_adapt/segment_adapt.pl"
+
+  # ラベルと raw ファイルを HTS-demo_NIT-ATR503-M001 内に配置
+  cp $STARTPATH/tools/segment_adapt/voices/data/raw/*.raw $STARTPATH/tools/HTS-demo_NIT-ATR503-M001/data/raw/
+  cp $STARTPATH/tools/segment_adapt/voices/data/full/*.lab $STARTPATH/tools/HTS-demo_NIT-ATR503-M001/data/labels/full/
+  cp $STARTPATH/tools/segment_adapt/voices/data/mono/*.lab $STARTPATH/tools/HTS-demo_NIT-ATR503-M001/data/labels/mono/
+
 # splitAndGetLabel でラベルデータを作成
-echo -e "\nExecute splitAndGetLabel ...\n"
-if [ ${APIMODE} = "J" ] ; then
+elif [ ${APIMODE} = "J" ] ; then
+  echo -e "\nExecute splitAndGetLabel ...\n"
   ./splitAndGetLabel ${OPTIONS}
 else
+  echo -e "\nExecute splitAndGetLabel ...\n"
   ./splitAndGetLabel -${APIMODE} ${APIKEY} ${OPTIONS}
 fi
 
 # 音響モデルのビルド
-echo -e "Build acoustic model ...\n"
+echo -e "Build acoustic model ..."
 
-cd ../../HTS-demo_NIT-ATR503-M001/
+cd $STARTPATH/tools/HTS-demo_NIT-ATR503-M001/
 fileCount=`find data/raw -type f | wc -l`
+echo -e "File count: ${fileCount}\n"
 
 if [ ${fileCount} -lt 503 ] ; then
   # ファイル数が 503 以下 (ファイル数が少なくなればなるほどトレーニングの反復回数が増える)
@@ -113,16 +157,36 @@ make clean
 make
 
 # プロセスが実行されている間待機する
-# 出力内容は tail -f tools/HTS-demo_NIT-ATR503-M001/log で確認できる
+# 出力内容は tail -f tools/HTS-demo_NIT-ATR503-M001/log でも確認できる
+# 強制終了したいときは ps aux | grep Training.pl | grep -v grep | awk '{ print "kill -9", $2 }' | sh を実行
+
+# 初期化
 line=`ps x | grep Training.pl | grep -v grep`
+lineCount=0
+
+# プロセスが動いていれば
 while [ "$line" != "" ]
 do
-  sleep 1s
+  # 0.01 秒スリープ
+  sleep 0.01s
+  
+  # プロセスの実行状態を取得
   line=`ps x | grep Training.pl | grep -v grep`
+
+  # 更新されたログを随時表示
+  lineCount_new=`cat log | wc -l`
+  if [ `expr ${lineCount_new}` -gt `expr ${lineCount}` ]; then
+    log=`cat log | head -${lineCount_new} | tail`
+    echo -e "${log}" # まだ表示していない行だけ出力
+  fi
+  lineCount=`expr ${lineCount_new}`
 done
 
 # 生成された音響モデルを指定フォルダにコピー
 cd "$STARTPATH"
-cp tools/HTS-demo_NIT-ATR503-M001/voices/qst001/ver1/nitech_jp_atr503_m001.htsvoice "$HTSVOICEPATH"
-
-echo -e "Output to $HTSVOICEPATH. Finished.\n'
+if [ -e tools/HTS-demo_NIT-ATR503-M001/voices/qst001/ver1/nitech_jp_atr503_m001.htsvoice ]; then
+  cp tools/HTS-demo_NIT-ATR503-M001/voices/qst001/ver1/nitech_jp_atr503_m001.htsvoice "${HTSVOICEPATH}"
+  echo -e "\nOutput to \"${HTSVOICEPATH}\". Done.\n"
+else
+  echo -e "\nCouldn't Output to \"${HTSVOICEPATH}\". Failed.\n"
+fi
